fix: yarn.lock for frontend plugins
This change updates the CLI's frontend script to create a yarn.lock
based on the exported plugin package in a similar fashion to backend
plugins.  This makes the exported plugin package for frontend plugins
more consistent with backend plugins and allows another means
for security scanners to inspect the plugin's dependencies.  This change
also moves functions that are shared between the backend and frontend
commands into a shared utils file so it's more obvious which functions
are common to each command.